--- Career Endorsement Tracker/Career Endorsement Tracker/Models/Task.swift
@@ -1,14 +1,0 @@
-//
-//  Task.swift
-//  Career Endorsement Tracker
-//
-//  Created by Victor  on 8/27/19.
-//  Copyright © 2019 Lambda School. All rights reserved.
-//
-
-import Foundation
-
-struct Task {
-    var name: String
-    var description: String
-}
